+++ examples/advanced/src/gen/models/ts/Dog.ts
@@ -1,0 +1,11 @@
+export type Dog = {
+  /**
+   * @minLength 1
+   * @type string
+   */
+  readonly type: string
+  /**
+   * @type string
+   */
+  name: string
+}