+++ examples/zod/src/gen/zod/createUserSchema.gen.ts
@@ -9,14 +9,14 @@ import { userSchema } from './userSchema.gen.ts'
 /**
  * @description successful operation
  */
-export const createUserErrorSchema = userSchema
+export const createUserErrorSchema = z.lazy(() => userSchema)
 
 export type CreateUserErrorSchema = z.infer<typeof createUserErrorSchema>
 
 /**
  * @description Created user object
  */
-export const createUserMutationRequestSchema = userSchema
+export const createUserMutationRequestSchema = z.lazy(() => userSchema)
 
 export type CreateUserMutationRequestSchema = z.infer<typeof createUserMutationRequestSchema>