+++ examples/zod/src/gen/zod/customerSchema.gen.ts
@@ -7,11 +7,9 @@ import { z } from '../../zod.ts'
 import { addressSchema } from './addressSchema.gen.ts'
 
 export const customerSchema = z.object({
-  id: z.int().optional(),
+  id: z.number().int().optional(),
   username: z.string().optional(),
-  get address() {
-    return z.array(addressSchema).optional()
-  },
+  address: z.array(z.lazy(() => addressSchema)).optional(),
 })
 
 export type CustomerSchema = z.infer<typeof customerSchema>