+++ examples/zod/src/gen/zod/petSchema.gen.ts
@@ -8,19 +8,15 @@ import { categorySchema } from './categorySchema.gen.ts'
 import { tagSchema } from './tagSchema.gen.ts'
 
 export const petSchema = z.object({
-  id: z.int().optional(),
+  id: z.number().int().optional(),
   internalId: z
     .string()
     .regex(/^[0-9]{1,19}$/)
     .optional(),
   name: z.string().uuid(),
-  get category() {
-    return categorySchema.optional()
-  },
+  category: z.lazy(() => categorySchema).optional(),
   photoUrls: z.array(z.string()),
-  get tags() {
-    return z.array(tagSchema).optional()
-  },
+  tags: z.array(z.lazy(() => tagSchema)).optional(),
   status: z.enum(['available', 'pending', 'sold']).describe('pet status in the store').optional(),
 })