+++ examples/zod/src/gen/zod/updatePetSchema.gen.ts
@@ -9,7 +9,7 @@ import { petSchema } from './petSchema.gen.ts'
 /**
  * @description Successful operation
  */
-export const updatePet200Schema = petSchema
+export const updatePet200Schema = z.lazy(() => petSchema)
 
 export type UpdatePet200Schema = z.infer<typeof updatePet200Schema>
 
@@ -37,10 +37,10 @@ export type UpdatePet405Schema = z.infer<typeof updatePet405Schema>
 /**
  * @description Update an existent pet in the store
  */
-export const updatePetMutationRequestSchema = petSchema
+export const updatePetMutationRequestSchema = z.lazy(() => petSchema)
 
 export type UpdatePetMutationRequestSchema = z.infer<typeof updatePetMutationRequestSchema>
 
-export const updatePetMutationResponseSchema = updatePet200Schema
+export const updatePetMutationResponseSchema = z.lazy(() => updatePet200Schema)
 
 export type UpdatePetMutationResponseSchema = z.infer<typeof updatePetMutationResponseSchema>