+++ packages/plugin-zod/src/generators/__snapshots__/anyof.ts
@@ -5,15 +5,11 @@
 import { z } from 'zod'
 
 export const test = z.union([
-  z
-    .object({
-      propertyA: z.string(),
-    })
-    .strict(),
-  z
-    .object({
-      propertyA: z.string(),
-      propertyB: z.string(),
-    })
-    .strict(),
+  z.object({
+    propertyA: z.string(),
+  }),
+  z.object({
+    propertyA: z.string(),
+    propertyB: z.string(),
+  }),
 ])