+++ packages/plugin-svelte-query/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @kubb/plugin-svelte-query
 
+## 3.11.1
+
+### Patch Changes
+
+- Updated dependencies [[`5400e56`](https://github.com/kubb-labs/kubb/commit/5400e56fd866dbee721cd2dcbdb288088c58d990)]:
+  - @kubb/plugin-zod@3.11.1
+  - @kubb/plugin-client@3.11.1
+  - @kubb/core@3.11.1
+  - @kubb/oas@3.11.1
+  - @kubb/plugin-oas@3.11.1
+  - @kubb/plugin-ts@3.11.1
+  - @kubb/react@3.11.1
+
+## 3.11.0
+
+### Patch Changes
+
+- Updated dependencies [[`13189ee`](https://github.com/kubb-labs/kubb/commit/13189ee0c7b297cc42cf9a7d476780ff7e357efe), [`55de3d2`](https://github.com/kubb-labs/kubb/commit/55de3d2758ce4957882243ad70d3168d3c41ff40)]:
+  - @kubb/plugin-zod@3.11.0
+  - @kubb/plugin-oas@3.11.0
+  - @kubb/plugin-client@3.11.0
+  - @kubb/plugin-ts@3.11.0
+  - @kubb/core@3.11.0
+  - @kubb/oas@3.11.0
+  - @kubb/react@3.11.0
+
+## 3.10.15
+
+### Patch Changes
+
+- Updated dependencies [[`db73926`](https://github.com/kubb-labs/kubb/commit/db73926f46739e598244bedc52f466591b2d7320)]:
+  - @kubb/plugin-ts@3.10.15
+  - @kubb/plugin-client@3.10.15
+  - @kubb/plugin-zod@3.10.15
+  - @kubb/core@3.10.15
+  - @kubb/oas@3.10.15
+  - @kubb/plugin-oas@3.10.15
+  - @kubb/react@3.10.15
+
 ## 3.10.14
 
 ### Patch Changes